envoy: Initial Envoy integration.
This commit adds three Envoy-related subdirectories to Cilium repo:

envoy: Envoy proxy in a git submodule (in envoy/envoy) and cilium
       filters. Running tests verifies that Envoy binary was successfully
       built.

envoy-api: Envoy-api protobuf definitions in submodule
           envoy-api/envoy-api, and it's google dependency in submodule
           envoy-api/google. A small test program (test.go) is included to test
           launching of Envoy and operating it's LDS and RDS APIs.

pkg/envoy: Go packge for interfacing with Envoy, currently used only
           by test.go mentioned above.

This is still work-in-progress, but should be good enough for making
CI have all the required dependencies for successful builds.

Note that for time being the envoy and envoy-api dependencies point to
my forks on Github as we depend on additional functionality not yet
available in Envoy upstream.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>